+++ perun/logic/config.py
@@ -38,11 +38,13 @@ def init_shared_config_at(path):
     store.touch_file(path)
 
     shared_config = streams.safely_load_yaml_from_stream("""
-global:
-    profile_info_fmt: "\u2503 [type] \u2503 [collector]  \u2503 ([time]) \u2503 [origin] \u2503"
-    minor_version_info_fmt: "[id:6] ([stats]) [desc]"
+general:
     editor: vim
     paging: only-log
+
+format:
+    status: "\u2503 [type] \u2503 [collector]  \u2503 ([time]) \u2503 [origin] \u2503"
+    shortlog: "[id:6] ([stats]) [desc]"
     """)
 
     write_config_file(shared_config, path)